% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data.R
\name{dl_data}
\alias{dl_data}
\title{Download the 4 dataset to run chouetteprojet}
\usage{
dl_data(data_url, database, file_name)
}
\arguments{
\item{data_url}{A character string (or a vector of character strings): the url(s)
where data can be found.}

\item{database}{A character string (or a vector of character strings): the database(s)
where data is stored.}

\item{file_name}{A character string (or a vector of character strings): the name(s)
of the destination file(s) for the data that is downloaded.}
}
\value{
No return value.
}
\description{
Download the 4 dataset to run chouetteprojet
}
